requirements.txt
Pillow
chime
pyarrow
python-dotenv
requests
pytest
